Testing area  (#69)
* Delete CMakeCache.txt

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* Delete CMakeCCompiler.cmake

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* Delete cmake_install.cmake

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* Delete CMakeCXXCompiler.cmake

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* Delete CMakeSystem.cmake

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* Delete CMakeCCompilerId.c

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* Delete CXX.includecache

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* Delete progress.make

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* Delete CXX.includecache

ficheiros produzidos pelo cmake nao e' necessario unir ao master

* gitignore update to remove build directory

* gitignore update to remove build directory

* merge from master

* issue #62 seems to be solved. is was  a misnamer in the import of parameters

* master merge sync

* solving issue #62 at energy branch

* corrected typo on cmake tests

* code cleaning assisted by clang-tidy and clion

* typos and grammar revision on .md files

* added ini file mainly for testing purposes

* dummy initialization of grid relative points

* simplificaçao dos ciclos da leitura dos ficheiro ini

* static cast to clean code

* 2d test from ini .file

* added TODO directives

* .ini file can now have blanck intermediate lines, line comments with ';' and sections with [section]

* .ini file can now have blanck intermediate lines, line comments with ';' and sections with [section]

* implementation and testing of forced oscilations at the boundary layer to study pulsatile flow

+++ src/ElectricLib.cpp
@@ -1,6 +1,6 @@
 #include "includes/ElectricLib.h"
-#include "GrapheneFluid2DLib.h"
-#include "GrapheneFluid1DLib.h"
+#include "includes/GrapheneFluid2DLib.h"
+#include "includes/GrapheneFluid1DLib.h"
 
 #ifndef MAT_PI
 #    define MAT_PI 3.14159265358979323846
@@ -56,7 +56,7 @@ void ElectroAnalysis::ComputeElectroBase(float t, const GrapheneFluid2D& graphen
 
 void ElectroAnalysis::ComputeElectroDerived() {
 	float dt;
-	dt=TmpArr.back()/DipX.size();
+	dt=TmpArr.back()/static_cast<float>(DipX.size());
 	EngCap.resize(DipX.size());
 	PowCap.resize(DipX.size());
 	transform(NetQ.begin(), NetQ.end(), EngCap.begin(), [](const float &c){ return 0.5f*c*c; });
@@ -154,7 +154,6 @@ float ElectroAnalysis::OhmPower(const GrapheneFluid2D& graphene){
 		square_current_density[c] = (jx*jx+jy*jy)/sqrt(graphene.Den[c]);
 	}
 	return Integral_2_D(graphene.SizeX(), graphene.SizeY(), graphene.GetDx(), graphene.GetDy(), square_current_density);
-
 }
 
 float ElectroAnalysis::ElectricDipoleX(const GrapheneFluid2D &graphene) {
